# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Domain:
    properties:
      ICP_reg:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      vendor:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      domains:
        items:
          $ref: '#/definitions/models.Domain'
        type: array
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      stuId:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.UserRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      Admin: can submit new domain, access and delete all domains, promte/demote user
        to Contributor, grant/deny all domains access to other Normal user
      Contributor: can submit new domain, access and delete own domain, grant/deny
        own domain access to other Normal user
      Normal: only can access granted domains
      SysAdmin: same as Admin, promte/demote user to Admin
    x-enum-varnames:
    - Normal
    - Contributor
    - Admin
    - SysAdmin
...
  web.User:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: makiras.x@outlook.com
    name: domain0
  description: Domain0 API
  license:
    name: MPL(mozilla public license)-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
  title: Domain0 API
  version: 0.0.1
paths:
...
  /api/v1/user:
    get:
      consumes:
      - application/json
      description: |-
        Get user list
        Only admin can get user list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
      summary: Get user list
      tags:
      - user
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete user by id
        Only admin can delete user, and the user must has none role.
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
      summary: Delete user
      tags:
      - user
    get:
      description: |-
        Get user info by id
        if id is not the same as jwt sub, jwt role must be admin
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
      summary: Get user info
      tags:
      - user
    put:
      consumes:
      - application/json
      description: |-
        Update user info by id
        Only admin can update other user info whoes role is lower than his.
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
      summary: Update user info
      tags:
      - user
...
schemes:
- http
- https
swagger: "2.0"